fix: hot repos separate component
